move sync mirror actions to mirror service
