+++ services/mirror/mirror.go
@@ -329,7 +329,7 @@ func SyncMirrors() {
 
 			// Create reference
 			if result.oldCommitID == gitShortEmptySha {
-				if err = models.MirrorSyncCreateAction(m.Repo, result.refName); err != nil {
+				if err = MirrorSyncCreateAction(m.Repo, result.refName); err != nil {
 					log.Error("MirrorSyncCreateAction [repo_id: %d]: %v", m.RepoID, err)
 				}
 				continue
@@ -337,7 +337,7 @@ func SyncMirrors() {
 
 			// Delete reference
 			if result.newCommitID == gitShortEmptySha {
-				if err = models.MirrorSyncDeleteAction(m.Repo, result.refName); err != nil {
+				if err = MirrorSyncDeleteAction(m.Repo, result.refName); err != nil {
 					log.Error("MirrorSyncDeleteAction [repo_id: %d]: %v", m.RepoID, err)
 				}
 				continue
@@ -359,7 +359,7 @@ func SyncMirrors() {
 				log.Error("CommitsBetweenIDs [repo_id: %d, new_commit_id: %s, old_commit_id: %s]: %v", m.RepoID, newCommitID, oldCommitID, err)
 				continue
 			}
-			if err = models.MirrorSyncPushAction(m.Repo, models.MirrorSyncPushActionOptions{
+			if err = MirrorSyncPushAction(m.Repo, MirrorSyncPushActionOptions{
 				RefName:     result.refName,
 				OldCommitID: oldCommitID,
 				NewCommitID: newCommitID,
